GCP Projects/day_06/ex_00/requirements.txt
Flask==2.0.1
google-cloud-datastore==2.1.6
google-auth==2.0.0
requests==2.26.0
pdfkit==0.6.1
wkhtmltopdf==0.2
beautifulsoup4==4.10.0